backend/fastapi_service/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
tensorflow
numpy
pillow
google-generativeai
python-dotenv
PyMuPDF
